**ft(jwt-token):Users get a token** for every user who successfully registers, a token is given to them so does one who logs in with right credentials. [Finishes #160583993]
